# server_go/docs/swagger.yaml
basePath: /
definitions:
  dto.CategoryBrief:
    properties:
      created_at:
        description: '"2006-01-02 15:05:05"'
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        description: 同上
        type: string
    type: object
...
  dto.PostBrief:
    properties:
      cate_id:
        type: integer
      created_at:
        description: '"2006-01-02 15:05:05"'
        type: string
      id:
        type: integer
      title:
        type: string
      title_en:
        type: string
      updated_at:
        description: 同上
        type: string
      user_id:
        type: integer
    type: object
  dto.UserBrief:
    properties:
      created_at:
        description: '"2006-01-02 15:05:05"'
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        description: 同上
        type: string
    type: object
...
  req.GetPostListReq:
    properties:
      cate_id:
        type: integer
      limit:
        type: integer
      page:
        type: integer
    type: object
...
  resp.PostDetailResp:
    properties:
      cate_id:
        type: integer
      category:
        $ref: '#/definitions/dto.CategoryBrief'
      content:
        description: 文章详情
        type: string
      created_at:
        description: '"2006-01-02 15:05:05"'
        type: string
      id:
        type: integer
      title:
        type: string
      title_en:
        type: string
      updated_at:
        description: 同上
        type: string
      user:
        $ref: '#/definitions/dto.UserBrief'
      user_id:
        type: integer
    type: object
  resp.PostListResp:
    properties:
      post_list:
        items:
          $ref: '#/definitions/dto.PostBrief'
        type: array
      total:
        type: integer
    type: object
...
host: localhost:8030
info:
  contact: {}
  description: yonder博客的后端API服务
  termsOfService: https://github.com/swaggo/swag
  title: yonder blog api aservice
  version: "1.0"
paths:
...
  /api/v1/post/detail:
    get:
      consumes:
      - application/json
      description: 获取文章详情，包括用户和分类信息
      parameters:
      - enum:
        - html
        in: query
        name: content_type
        type: string
      - in: query
        name: post_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.PostDetailResp'
      summary: 文章详情
      tags:
      - post
  /api/v1/post/list:
    get:
      consumes:
      - application/json
      description: 获取文章列表，支持分页，可按分类id进行查询
      parameters:
      - description: 查询参数
        in: body
        name: object
        schema:
          $ref: '#/definitions/req.GetPostListReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.PostListResp'
      summary: 文章列表
      tags:
      - post
...
swagger: "2.0"